+++ NEWS.md
@@ -1,3 +1,9 @@
+# shade 0.1.1
+
+* Fixed bug where Shiny app allowed individual threshold values to be greater than the base
+
+* Added documentation links to the Shiny app
+
 # shade 0.1.0
 
 * Initial version